re -write core to work with modules
