+++ scripts/src/core/configuration.ts
@@ -1,11 +1,7 @@
-/// <reference path="../core/location.ts" />
+import * as location from "./location"
 
-"use strict";
+const SCRIPTS_BASE_URL: string = location.relative("scripts/bin");
 
-module fugazi.config {
-	var SCRIPTS_BASE_URL: string = location.relative("scripts/bin");
-
-	export function scriptUrl(path: string): string {
-		return SCRIPTS_BASE_URL + "/" + path;
-	}
+export function scriptUrl(path: string): string {
+	return SCRIPTS_BASE_URL + "/" + path;
 }
